Fix Analyser warning in LOTRadialGradientLayer
The static analyser (using Xcode 9.2) was complaining about `gradientComponents` being a variable length array with the possibility of having 0 length. This could occur with `numberOfComponents` keeping the value of 0 if no colours exist. Having a VLA with size 0 causes undefined behaviour.

The change is to bypass rendering the LOTRadialGradientLayer if there are no colours provided, as it doesn't make sense to render anything without a colour, and to presume a default colour may not be the right solution.

Former-commit-id: 16ef457087b550bd681d7fe8110daa7fda6298c1